+++ lib/witchcraft/monoid/identity.ex
@@ -1,0 +1,6 @@
+defmodule Witchcraft.Monoid.Identity do
+  @doc ~S"""
+  Get the identity without providing a sample value
+  """
+  @callback identity() :: any
+end
